+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Sanity check csv rows for disallowed values
+- Let 12300 be the last debitor account number
 
 ## [0.4.5] - 2023-03-18